coerce tran_id to numeric consistently across all transfer returns
A Binance transaction id is a large identifier that overflows R's 32-bit
integer, so it must land as a double. The spot/sub-account transfer
endpoints already coerced and typed it numeric; the margin (borrow,
repay, isolated transfer), futures (income history), and sub-account
(add_transfer, transfer history) endpoints left it uncoerced and typed
integer, which only held because their fixtures use in-range ids — a real
out-of-range id would flip the column to double and break the contract.

Coerce tran_id with as.numeric() in those parsers and type it (numeric);
update empty_dt_futures_income's zero-row column to numeric(0) to match.
Genuine whole-number counters stay integer.

+++ R/helpers_parse.R
@@ -1250,7 +1250,7 @@ empty_dt_futures_income <- function() {
     asset = character(0),
     info = character(0),
     time = ms_to_datetime(numeric(0)),
-    tran_id = integer(0),
+    tran_id = numeric(0),
     trade_id = character(0)
   ))
 }
